+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [10.6.0] (melonJS 2) - _2022-xx-xx_
 
+### Added
+- Renderable : add the possibility to specify a blend mode per renderable (still limited to normal or multiply)
+
 ### Changed
 - Renderable : refactor of Text and BitmapText to isolate metrics related features/functions to standardize the api between the 2 classes